Move the HelloGalaxy Maven example to checker/examples/MavenExample, add it to the release, and document it. Deprecate (but don't yet remove) the javac_maven script and javac_maven.bat.
