+++ checker/bin/README
@@ -1,3 +1,7 @@
+NOTE: javac_maven and javac_maven.bat are deprecated and will be removed
+in the December 1, 2015 release.
+
+
 This directory, "checker/bin", contains scripts to run the Checker Framework.
 
 javac - Is a shell script that runs the Checker Framework in *nix systems
@@ -7,15 +11,6 @@ This script is a drop-in replacement for the script javac provided by the
 OpenJDK.
 
 javac.bat - Is the equivalent of the javac script for Windows systems.
-
-javac_maven - Similar to the javac script, but meant for use with Maven
-pom files.  It requires a file name 'argfile' to be in the same directory
-as the pom file - this file must contain the command-line arguments
-to be passed to javac indicating the checker to be used and optional
-parameters to the Checker Framework.
-
-javac_maven.bat - Is the equivalent of the javac_maven script for Windows
-systems.
 
 
 The directory "checker/dist/" houses all jars built for the binary distribution.
