+++ checker/examples/MavenExample/README
@@ -1,0 +1,15 @@
+This directory contains an example Maven project that downloads the latest
+Checker Framework Maven artifacts and runs the Nullness Checker.
+The project assumes you are using a Java 7 compiler.  If you are using
+a Java 8 compiler, modify the pom.xml by following the instructions in
+the three locations in the pom.xml that mention "if using Java 8".
+
+To run the example, simply execute:
+mvn compile
+or
+mvn package
+
+You should see output that includes:
+
+MavenExample.java:[28,29] error: [assignment.type.incompatible] incompatible types in assignment.
+
