+++ checker/examples/MavenExample/src/main/java/org/checkerframework/example/MavenExample.java
@@ -1,4 +1,4 @@
-package org.checkerframework.galaxy;
+package org.checkerframework.example;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -8,22 +8,20 @@ import org.apache.commons.lang3.text.StrBuilder;
 
 /**
  * If you run:
- * maven checkers:check
+ * mvn compile
  * or
- * maven package
+ * mvn package
  *
  * The build for this project should fail with a warning for the line:
  * @NonNull Object nn = nullable;
  *
- * I have included commons lang3 to demonstrate that the Checker Framework Maven plugin
- * will indeed find the correct Maven classpath.
  */
-public class HelloGalaxy {
+public class MavenExample {
 
     public static @Nullable Object nullable = null;
 
     public static void main(final String[] args) {
-        System.out.println("Hello Galaxy!");
+        System.out.println("Hello World!");
 
         StrBuilder stb = new StrBuilder();
