+++ checker/manual/external-tools.tex
@@ -308,6 +308,10 @@ instructions below.
 These instructions use the artifacts from
 \href{http://search.maven.org/#search\%7Cga\%7C1\%7Corg.checkerframework}{Maven Central}.
 
+See the demonstration \code{examples/MavenExample/} for an example of the use of
+these instructions in a Maven project.  This example can be used to verify that
+the Maven artifacts are retrieved correctly.
+
 \begin{enumerate}
 
 \item Declare a dependency on the Checker Framework artifacts.  Find the
