+++ checker/manual/introduction.tex
@@ -677,7 +677,7 @@ Progress tracing; see Section~\ref{debugging-options-progress}
 \item
 \<-AoutputArgsToFile>
 Output the compiler command-line arguments to a file.  Useful when this is not
-fully in your control, such as when the Checker Framework is run from Maven.
+fully in your control.
 See Section~\ref{debugging-options-output-args}
 
 \item
