--- maven-plugin/mvn-examples/HelloGalaxy/pom.xml
@@ -1,119 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>org.checkerframework</groupId>
-  <artifactId>MvnExample</artifactId>
-  <version>1.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-  <name>MvnExample</name>
-  <url>http://maven.apache.org</url>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.checkerframework</groupId>
-      <artifactId>checker-qual</artifactId>
-      <version><!-- checker-framework-version -->1.9.7<!-- /checker-framework-version --></version>
-    </dependency>
-
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-lang3</artifactId>
-      <version>3.1</version>
-    </dependency>
-
-    <dependency>
-        <groupId>javax.servlet</groupId>
-        <artifactId>servlet-api</artifactId>
-        <version>2.4</version>
-        <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.checkerframework</groupId>
-        <artifactId>checkerframework-maven-plugin</artifactId>
-        <version><!-- checker-framework-version -->1.9.7<!-- /checker-framework-version --></version>
-        <executions>
-           <execution>
-             <!-- run the checkers after compilation; this can also be any later phase -->
-             <phase>process-classes</phase>
-             <goals>
-                 <goal>check</goal>
-             </goals>
-           </execution>
-        </executions>
-        <configuration>
-          <!-- required configuration options -->
-          <!-- a list of processors to run -->
-          <processors>
-              <processor>org.checkerframework.checker.nullness.NullnessChecker</processor>
-          </processors>
-
-            <checkerFrameworkVersion><!-- checker-framework-version -->1.9.7<!-- /checker-framework-version --></checkerFrameworkVersion>
-
-          <!-- setting procOnly to false will result in actual generation of class files -->
-          <procOnly>false</procOnly>
-          <!-- other optional configuration -->
-          <!-- full path to a java executable that should be used to create the forked JVM -->
-          <executable>/usr/bin/java</executable>
-          <!-- should an error reported by a checker cause a build failure, or only be logged as a warning; defaults to true -->
-          <failOnError>false</failOnError>
-
-
-          <procOnly>false</procOnly>
-          <!-- additional parameters passed to the JSR 308 Java compiler -->
-          <!--<javacParams>-Awarns -Xmaxwarns 10000</javacParams>-->
-          <javacParams>-Alint</javacParams>
-          <!-- additional parameters to pass to the forked JVM -->
-          <javaParams>-Xdebug</javaParams>
-
-        </configuration>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-site-plugin</artifactId>
-        <version>3.2</version>
-      </plugin>
-    </plugins>
-  </build>
-
-    <reporting>
-        <excludeDefaults>true</excludeDefaults>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jxr-plugin</artifactId>
-                <version>2.1</version>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>findbugs-maven-plugin</artifactId>
-                <!--version>2.2</version-->
-                <version>2.0.1</version>
-                <configuration>
-                    <findbugsXmlOutput>true</findbugsXmlOutput>
-                    <findbugsXmlWithMessages>true</findbugsXmlWithMessages>
-                    <includeTests>true</includeTests>
-                    <xmlOutput>true</xmlOutput>
-                </configuration>
-            </plugin>
-        </plugins>
-    </reporting>
-
-</project>
