+++ release/sanity_checks.py
@@ -79,7 +79,7 @@ def javac_sanity_check( checker_framework_website, release_version ):
 def maven_sanity_check( sub_sanity_dir_name, repo_url, release_version ):
     """
        Download the Checker Framework maven plugin from the given repository.  Download the
-       HelloGalaxy example for the Maven plugin and run the NullnessChecker on it.  If we don't
+       MavenExample example for the Maven plugin and run the NullnessChecker on it.  If we don't
        encounter the expected errors fail.
     """
     checker_dir = os.path.join(CHECKER_FRAMEWORK, "checker")
@@ -102,20 +102,20 @@ def maven_sanity_check( sub_sanity_dir_name, repo_url, release_version ):
         if os.path.isdir( path_to_artifacts ):
             delete_path( path_to_artifacts )
 
-        hello_galaxy_dir = os.path.join( maven_sanity_dir, "HelloGalaxy" )
-        output_log  = os.path.join( hello_galaxy_dir, "output.log" )
+        maven_example_dir = os.path.join( maven_sanity_dir, "MavenExample" )
+        output_log  = os.path.join( maven_example_dir, "output.log" )
 
         ant_release_script = os.path.join( CHECKER_FRAMEWORK_RELEASE, "release.xml" )
-        get_example_dir_cmd = "ant -f %s update-and-copy-hello-galaxy -Dchecker=%s -Dversion=%s -Ddest.dir=%s" % ( ant_release_script, checker_dir, release_version, maven_sanity_dir )
+        get_example_dir_cmd = "ant -f %s update-and-copy-maven-example -Dchecker=%s -Dversion=%s -Ddest.dir=%s" % ( ant_release_script, checker_dir, release_version, maven_sanity_dir )
 
         execute( get_example_dir_cmd )
 
-        hello_galaxy_pom = os.path.join( hello_galaxy_dir, "pom.xml" )
-        add_repo_information( hello_galaxy_pom, repo_url )
+        maven_example_pom = os.path.join( maven_example_dir, "pom.xml" )
+        add_repo_information( maven_example_pom, repo_url )
 
-        execute_write_to_file( "mvn checkerframework:check", output_log, False, hello_galaxy_dir )
+        execute_write_to_file( "mvn compile", output_log, False, maven_example_dir )
         check_results( "Maven sanity check", output_log, [
-            "HelloGalaxy.java:[30,29] [assignment.type.incompatible] incompatible types in assignment."
+            "MavenExample.java:[28,29] [assignment.type.incompatible] incompatible types in assignment."
         ])
 
         delete_path( path_to_artifacts )
